% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff38_label}
\alias{ff38_label}
\title{Convert Fama-French 38 industry codes to labels}
\usage{
ff38_label(ff38)
}
\arguments{
\item{ff38}{A numeric vector of fama-french industry codes}
}
\value{
A vector of Fama-French 38 industry portfolio labels
}
\description{
Converts a vector of Fama-French industry codes to the appropriate labels
}
\examples{
x <- c(1,14,37)
ff38_label(x)
}
