% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indclass.r
\docType{package}
\name{indclass}
\alias{indclass}
\alias{indclass-package}
\title{indclass: a collection of tools for working with industry codes}
\description{
indclass is supposed to make working with industry codes simpler and faster.
}
