% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naics_group}
\alias{naics_group}
\title{Convert full NAICS codes to 2, 3, 4, or 5 digit codes}
\usage{
naics_group(NAICS, digit)
}
\arguments{
\item{NAICS}{A numeric vector of NAICS codes}

\item{digit}{How many digits to shorten the NAICS code by}
}
\value{
A numeric vector of shortened NAICS industry codes
}
\description{
Converts NAICS industry classifications into their broader categories by
shortening the 6 digit code to 2, 3, 4, or 5 digits
}
\examples{
x <- c(800000,205000,190000)
naics_group(x, 3)
}
