% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sic_to_ff38}
\alias{sic_to_ff38}
\title{Convert SIC codes to Fama French 38 industry codes}
\usage{
sic_to_ff38(SIC)
}
\arguments{
\item{SIC}{A numeric vector of SIC codes}
}
\value{
A numeric vector of Fama-French 38 industry portfolio codes
}
\description{
Converts SIC codes to their corresponding industry code using the Fama-French
38 industry portfolio classifications
}
\examples{
x <- c(800,2000,4537)
sic_to_ff38(x)
}
